% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualMean.R
\name{CalculateAnnualMean}
\alias{CalculateAnnualMean}
\title{Calculates annual mean}
\usage{
CalculateAnnualMean(dt, method = "WMO")
}
\arguments{
\item{dt}{data frame}

\item{method}{WMO or normal}
}
\description{
Calculates annual mean
}
