% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{ComputeDailyClimatology}
\alias{ComputeDailyClimatology}
\title{Compute daily climatology}
\usage{
ComputeDailyClimatology(data, startYear, endYear)
}
\arguments{
\item{data}{data.table daily data}

\item{startYear}{int starting year}

\item{endYear}{int end year}
}
\description{
Compute daily climatology
}
