% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{Hindcast}
\alias{Hindcast}
\title{Hindcast}
\usage{
Hindcast(year, dailyData, dailyClimatology)
}
\arguments{
\item{year}{integer year for hindcast}

\item{dailyData}{of hindcast year}

\item{dailyClimatology}{daily climatology}
}
\description{
Necessary to compute, bias, sd, rmse and other statistics of projection
}
