% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{HindcastProjection}
\alias{HindcastProjection}
\title{Hindcast}
\usage{
HindcastProjection(year, dailyData, dailyClimatology)
}
\arguments{
\item{year}{integer year for hindcast}

\item{dailyClimatology}{daily climatology}

\item{daily}{data of hindcast year}
}
\description{
Necessary to compute, bias, sd, rmse and other statistics of projection
}
