% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKISData.R
\name{KIS}
\alias{KIS}
\title{Extract data from KIS}
\usage{
KIS(var, geoIdentifier, period)
}
\arguments{
\item{var}{Variable to extract (plural?)}

\item{geoIdentifier}{Station identifier (, spatial point, spatial area ..., plural?)}

\item{period}{Xts}
}
\value{
data.table
}
\description{
Extract data from KIS
}

