% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractKISData.R
\name{KIStemplate}
\alias{KIStemplate}
\title{Python test template}
\usage{
KIStemplate(var, geoIdentifier, period)
}
\arguments{
\item{var}{variable}

\item{geoIdentifier}{location}

\item{period}{period}
}
\description{
Python test template
}
