% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{MakeTrendEnvelope}
\alias{MakeTrendEnvelope}
\title{Produces trend envelope data frame}
\usage{
MakeTrendEnvelope(start, end)
}
\arguments{
\item{start}{dt with year, lower, and upper}

\item{end}{dt with year, lower, and upper}
}
\description{
uses linear interpolation
}
