% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{MeanProjection}
\alias{MeanProjection}
\title{Projection}
\usage{
MeanProjection(day, measurements, forecast, climatology, statistics,
  sdfactor = 1.96)
}
\arguments{
\item{day}{date of projection}

\item{measurements}{daily data}

\item{forecast}{14-day forecast (not implemented yet)}

\item{climatology}{daily climatology}

\item{statistics}{statistics for bias correction and uncertainty}

\item{sdfactor}{factor to determine cofidence interval (default = 1.96)}
}
\description{
Projection
}
