% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{PredictGamlss}
\alias{PredictGamlss}
\title{Gamlss projection}
\usage{
PredictGamlss(Date, dt, forecast, probs = c(0.05, 0.5, 0.95))
}
\arguments{
\item{Date}{date from which to predict}

\item{dt}{data.table with daily measurements (at least last 30 years)}

\item{forecast}{data.table with operational forecast}

\item{probs}{probabilities to predict}
}
\description{
Gamlss projection
}
