% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{data}
\name{tmp}
\alias{tmp}
\title{tmp}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 42475 rows and 4 columns.}
\usage{
tmp
}
\description{
An example temperature data set
}
\keyword{datasets}
