% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{PredictGamlss}
\alias{PredictGamlss}
\title{Gamlss projection}
\usage{
PredictGamlss(Date, dt, forecast, probs = c(0.05, 0.5, 0.95))
}
\arguments{
\item{forecast}{14-day forecast (not implemented yet)}
}
\description{
Gamlss projection
}
