% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{PredictMovingWindow}
\alias{PredictMovingWindow}
\title{Moving window prediction}
\usage{
PredictMovingWindow(Date, dt, forecast = NULL, probs = c(0.05, 0.5, 0.95),
  k = 12)
}
\arguments{
\item{forecast}{14-day forecast (not implemented yet)}
}
\description{
Moving window prediction
}
